azure-functions/requirements.txt
# Azure Functions Core
azure-functions==1.18.0
azure-functions-worker==1.0.0

# Azure SDK Components
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-servicebus==7.11.0
azure-eventhub==5.11.4
azure-cosmos==4.5.1
azure-monitor-opentelemetry==1.2.0

# Database Connectivity
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
alembic==1.13.1

# Blockchain Integration
web3==6.15.1
eth-account==0.10.0
eth-hash==0.6.0
eth-typing==4.0.0

# HTTP and API
requests==2.31.0
aiohttp==3.9.1
fastapi==0.108.0
uvicorn==0.25.0

# Data Processing and Analytics
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2

# Cryptography and Security
cryptography==41.0.8
pyjwt[crypto]==2.8.0
bcrypt==4.1.2
argon2-cffi==23.1.0

# Monitoring and Logging
opencensus-ext-azure==1.0.8
opencensus-ext-logging==0.1.1
opencensus-ext-requests==0.8.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
marshmallow==3.20.2
celery==5.3.4
redis==5.0.1

# JSON and Serialization
orjson==3.9.12
ujson==5.9.0

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Validation and Parsing
cerberus==1.3.5
voluptuous==0.14.2

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-mock==3.12.0
faker==21.0.0

# Compliance and Audit
audit-python-lib==3.1.0

# Machine Learning (for fraud detection)
tensorflow==2.15.0
torch==2.1.2
xgboost==2.0.2

# Development Tools
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0